完成了 rsa 算法的附加返回字段功能，待测试。

+++ fusion-service/src/main/java/com/welab/fusion/service/service/JobService.java
@@ -225,8 +225,6 @@ public class JobService extends AbstractService {
     }
 
 
-
-
     /**
      * 保存任务成员
      *
@@ -392,11 +390,9 @@ public class JobService extends AbstractService {
         }
 
         // 删除求交结果
-        if (StringUtil.isNotEmpty(job.getResultFilePath())) {
-            File file = new File(job.getResultFilePath());
-            if (file.exists()) {
-                file.delete();
-            }
+        File file = job.getResultFile();
+        if (file != null && file.exists()) {
+            file.delete();
         }
 
         // 应删尽删
